openaidemo/pyproject.toml
[project]
name = "openaidemo"
version = "0.1.0"
description = "Demo project for trying out openai"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.97.1",
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
    "six>=1.17.0",
]
